backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
haystack-ai
sentence-transformers
python-multipart
python-dotenv
anthropic
